requirements.txt
annotated-types==0.7.0
anyio==4.9.0
backoff==2.2.1
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
distro==1.9.0
fastapi>=0.109.0
grpcio==1.71.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx>=0.26.0
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
jiter==0.9.0
langchain-core>=0.1.17
langchain-openai>=0.0.5
mem0ai>=0.1.98
monotonic==1.6
numpy==2.2.5
openai==1.78.1
packaging==24.0
pluggy==1.5.0
portalocker==2.10.1
posthog==3.25.0
protobuf==6.30.2
pydantic>=2.6.0
pydantic_core==2.33.2
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv>=1.0.0
pytz==2024.2
PyYAML==6.0.2
qdrant-client==1.14.2
requests==2.32.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn>=0.27.0
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
